+++ apps/s3dfm/event.c
@@ -124,7 +124,7 @@ int event_click(struct s3d_evt *evt)
 	int oid;
 	t_node *f;
 	s3dw_handle_click(evt);
-	oid=(int)*((unsigned long *)evt->buf);
+	oid=(int)*((u_int32_t *)evt->buf);
 	if (NULL!=(f=node_getbyoid(&root,oid)))
 	{
 		if (f->objs.close==oid)
